+++ ScrollComponent.h
@@ -198,9 +198,9 @@ private:
     ScrollItemCallback m_OnBindItem;
 
     // Input handlers (registered on InputCollider in Start())
-    void HandlePointerDown(int32_t x, int32_t y);
-    void HandlePointerMove(int32_t x, int32_t y);
-    void HandlePointerUp(int32_t x, int32_t y);
+    void HandlePointerDown(float x, float y);
+    void HandlePointerMove(float x, float y);
+    void HandlePointerUp(float x, float y);
 
     // Helpers
     int32_t MeasureChildSize(DekiObject* child) const;
